rename date module
